refactor: 아키텍처 구조 개선 - controllers/renderers를 ui/로 통합
