--- src/basic/controllers/index.js
@@ -1,14 +1,0 @@
-// ==================== 컨트롤러 레이어 통합 export ====================
-
-// 장바구니 컨트롤러 - 오케스트레이션 로직
-export {
-  updateSelectOptions,
-  calculateCartStuff,
-  renderBonusPoints,
-  updatePricesInCart,
-} from './CartController.js';
-
-// 장바구니 이벤트 컨트롤러 - 이벤트 처리 로직
-export {
-  setupCartEventListeners,
-} from './CartEventController.js';
