+++ src/basic/utils/index.js
@@ -1,4 +1,1 @@
-// ==================== 유틸리티 함수 통합 export ====================
-
-// 포맷터 유틸리티
 export { formatPrice, formatRoundedPrice } from './formatters.js';
